Moving components to separate files.
