+++ components/MyInfo.js
@@ -1,0 +1,18 @@
+import React from "react"
+
+// MyFunction
+function MyInfo() {
+    return (
+      <div>
+        <h1>Anshul Jain</h1>
+        <p>I am a backend python developer looking to learn React. My Top 3 Vacation Spots:</p>
+        <ol>
+          <li>Seattle, WA</li>
+          <li>Portland, OR</li>
+          <li>Beaverton, OR</li>
+        </ol>
+      </div>
+    )
+  }
+  
+export default MyInfo
