+++ index.js
@@ -10,21 +10,7 @@
 
 import React from "react"
 import ReactDOM from "react-dom"
-
-// MyFunction
-function MyInfo() {
-  return (
-    <div>
-      <h1>Anshul Jain</h1>
-      <p>I am a backend python developer looking to learn React. My Top 3 Vacation Spots:</p>
-      <ol>
-        <li>Seattle, WA</li>
-        <li>Portland, OR</li>
-        <li>Beaverton, OR</li>
-      </ol>
-    </div>
-  )
-}
+import MyInfo from "./components/MyInfo"
 
 // JSX
 ReactDOM.render(<MyInfo />, document.getElementById("root"))
